graph-tool loader now prints graph to stdout
